Add helper functions, improve FPR calc
Added helper functions for easier insertion and removal of keys.
Made the FPR calculation more efficient.
Added an internal count of entries, which allows for an estimation
of the current FPR of the bloom filter (read the docstring of the
size() function to find out about its limitation).
Added unittests for these new functions.