Fix problem with deletion routine
The deletion routine had an issue: When deleting keys, all
operations were performed instantly.  This means that, if a
key was removed that was never added to the bloom filter, and
if this problem was detected only with the second hash function,
the decrement of the first hash function would have already
happened, leaving the bloom filter in an inconsistent state.

This problem was fixed by deferring the operations until the
loop over the hash functions has terminated.  This way, the VICBF
will never be in an inconsistent state.