Change serialization format again
Serialization of individual key-value-pairs was potentially using
slot identifiers that weren't multiples of 8.  This is not a problem
when using python, but the Java counterpart has no ability to read
arbitrary numbers of bits.  Thus, even if it is less efficient, the
serialization will now use multiples of 8 bits during serialization.